refactor: switch to compound components
