+++ src/pages/Calculator.tsx
@@ -1,10 +1,9 @@
 import { IonAvatar, IonButton, IonInput, IonRippleEffect } from '@ionic/react';
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
-import { useAuth } from '../firebase/AuthContext';
 import { IModalProps, Modal } from '../components/Modal';
-import PageTitle from '../components/PageTitle';
 import Page from '../components/Page';
+import { useAuth } from '../firebase/AuthContext';
 
 const ICON_URL = `${process.env.PUBLIC_URL}/assets/icon/logo-circle.png`;
 
@@ -317,7 +316,7 @@ class Calculator extends React.Component<unknown, ICalculatorState> {
       <Page>
         <Outer>
           <div>
-            <PageTitle
+            <Page.Title
               title="Grade Predictor"
               subtitle="Tap the green chat bubbles to configure."
             />
